+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Hooch clash".
 - Tests for "Lunar lockout".
 - Tests for "Shikaku solver".
+- Tests for "Hexagonal maze".
 
 ### Changed
 - Renaming "Linear Bézier curves" to "Cubic Bézier curves".